Add testing

+++ Implementation/data_structures/htrie/htrie.h
@@ -20,22 +20,18 @@ namespace htrie
         /**
          * We store the complete trie in a single hash-map.
          *
-         * The key is a 32 bit integer where the first 28 bits (0th up to 27th)
-         * represent the current node index and the last 4 bits represent the
-         * current symbol used (this means the trie implementation has a span of
-         * 4).
+         * The key is a 32 bit integer where the first 24 bits (0th up to 23th)
+         * represent the current node index and the last 8 bits represent the
+         * current symbol used (this means the trie implementation has a span of 8).
          *
-         * We can differentiate a maximum of 2^28 or approx. 268 million
-         * different nodes which is enough for storing 32 bit integers.
+         * We can differentiate a maximum of 2^24 or approx. 16.7 million
+         * different nodes which is enough for storing all possible 32 bit integer values.
          *
-         * The value is again a 32 bit integer with the first 28 bits
-         * representing the node index this key is referring to.
+         * Since the trie will have height 3 (the last byte indicated stored using a special index) we will need a maximum of 256^3-1
          *
-         * Note that since we're only storing exactly 32bit keys with a static
-         * span of 4 we don't need to mark nodes as leaves as every node at the
-         * 4th layer must be a leaf indicating the presence of the given 32bit
-         * value.
+         * The value is again a 32 bit integer with either the first 24 bits
+         * representing the node index this key is referring to or .
          */
-        std::unordered_map<uint32_t, uint32_t> node_map_;
+        std::unordered_map<uint64_t, uint32_t> node_map_;
     };
 }
